+++ base/src/main/java/proguard/evaluation/BasicInvocationUnit.java
@@ -1,7 +1,7 @@
 /*
  * ProGuardCORE -- library to process Java bytecode.
  *
- * Copyright (c) 2002-2020 Guardsquare NV
+ * Copyright (c) 2002-2023 Guardsquare NV
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -241,7 +241,8 @@ implements   InvocationUnit,
         }
     }
 
-    private boolean mayBeExtension(Clazz clazz) {
+    protected boolean mayBeExtension(Clazz clazz)
+    {
         return clazz == null || (clazz.getAccessFlags() & FINAL) == 0;
     }
 }